Fix test failure on postgres

+++ src/pretix/base/models/orders.py
@@ -1691,8 +1691,9 @@ class OrderPosition(AbstractPosition):
         # Delete afterwards. Deleting in between might cause deletion of things related to add-ons
         # due to the deletion cascade.
         for cartpos in cp:
-            cartpos.addons.all().delete()
-            cartpos.delete()
+            if cartpos.pk:
+                cartpos.addons.all().delete()
+                cartpos.delete()
         return ops
 
     def __str__(self):
